requirements.txt

celery
django-celery==3.1.17
Django==1.9.7
django-bootstrap3==7.0.1
Pillow==3.4.2
gunicorn==19.6.0
djangorestframework==3.3.3
flower